Thursday Changes- Further PID Turning Tuning
-- PID turning tuned further (90+ done, <90 almost there)
-- ending delay increased to 30 on turns
-- increased tolerance on turns
-- changed PIDReturn for PIDCalc to use doubles instead of ints

+++ include/main.h
@@ -64,8 +64,8 @@ extern "C" {
 #endif
 // structures
 struct PIDReturn {
-    int prevError;
-    int prevIntegral; 
+    double prevError;
+    double prevIntegral; 
     int power;
 };
 struct ConstantContainer {
